L16 数据结构与算法/wxhh/requirements.txt
affinity==0.1.0
appdirs==1.4.2
attrs==16.3.0
autobahn==17.9.2
Automat==0.5.0
constantly==15.1.0
DBUtils==1.2
firefly==1.3.3.dev0
incremental==16.10.1
MySQL-python==1.2.5
packaging==16.8
pycrypto==2.6.1
pyparsing==2.1.10
python-memcached==1.58
six==1.10.0
SQLAlchemy==1.1.14
Twisted==17.1.0
txaio==2.8.2
ujson==1.35
zope.interface==4.3.3
